requirements.txt
amqp==2.6.1
asgiref==3.4.1
attrs==21.2.0
Automat==20.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==4.4.7
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
compress-pickle==2.0.1
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
DateTime==4.3
dj-database-url==0.5.0
Django==3.2
django-admin-interface==0.16.3
django-colorfield==0.4.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-json-widget==1.1.1
flower==0.9.5
future==0.18.2
fuzzywuzzy==0.18.0
h2==3.2.0
hpack==3.0.0
humanize==3.10.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.3.0
itemloaders==1.0.4
jmespath==0.10.0
kombu==4.6.11
lxml==4.6.3
lz4==3.1.3
numpy==1.21.0
pandas==1.3.0
parsel==1.6.0
priority==1.3.0
prometheus-client==0.8.0
Protego==0.1.16
psycopg2-binary==2.8.6
pyArango==1.3.5
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-Levenshtein==0.12.2
pytz==2021.1
queuelib==1.6.1
redis==3.5.3
requests==2.25.0
Scrapy==2.5.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tornado==6.1
typing-extensions==3.10.0.0
underscore.py==0.1.6
urllib3==1.26.6
vine==1.3.0
w3lib==1.22.0
whitenoise==5.2.0
zope.interface==5.4.0
